backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn==1.3.2  # Versi spesifik untuk konsistensi model
pandas
joblib
google-generativeai
python-dotenv
redis
sse_starlette